backend/requirements.txt
asgiref==3.11.1
attrs==25.3.0
beautifulsoup4==4.13.5
behave==1.2.6
black==26.3.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
cssselect==1.3.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
execnet==2.1.1
fasteners==0.20
filelock==3.19.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mycdp==1.2.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.6
pathspec==1.0.4
pdbp==1.7.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.11.0
pynose==1.5.5
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.4.2
pytest-django==4.12.0
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.0.1
pytest-xdist==3.8.0
python-dotenv==1.1.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruff==0.15.5
sbvirtualdisplay==1.4.0
selenium==4.35.0
seleniumbase==4.41.7
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.5
tabcompleter==1.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.3
undetected-chromedriver==3.5.5
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
